requirements.txt
# OpenAI API
openai>=1.0.0

# Common utilities (add more as needed)
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.28.0
requests>=2.31.0
elevenlabs>=1.10.0

# Development tools (optional but recommended)
# Uncomment if needed:
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
fal-client
moviepy>=1.0.3
pillow>=10.0.0
